Add files via upload
Today, as we are living in uncertainty, what may happen tomorrow no one knows,
whether it is related to our health or the objects around us that we use in our daily
life. So, why not take Insurance. Insurance is all about providing financial safety
to the individual. It helps us to enjoy financial security. If we have a proper
application to access all the Insurance related activity then it will be cheery on
top
to manage Insurance-related activity on a single platform. The admin is the user
who can check the list of the insurance holder, can add a category, subcategory,
add a policy, and etc. Another role here is that of the user who will be a customer
who can view the category, subcategory, policy and can apply for a policy listed
therein.
This project has all the necessary functionality that an Insurance management
application should have. It is built to support all roles, whether it is policymaker
and customer.
Insurance Management System where there are majorly two roles: One is the
admin and another one is the customer. Admin who is the main user of this
application will add the policy, add the category of Insurance along with the
subcategory. Another role is that of customers who will buy the policy from this
platform and can view the list of policies he/she holds.
The following are the major objective of this application:
1. To provide a bug-free application to the policymaker(admin) as well as
the customer.
2. The main objective is to build a secured, robust Insurance Management
system where the policies are managed properly.
3. It maintains the record of customers, policies, buyers’ policies
efficiently so that it would be easy to access at any time 24*7.
Modules and Functionalities of an Insurance Management
System
There are two main users of this application. One is the Admin whose
responsibility is to manage categories and policies and another role is that of
the customer who will buy the policy.
1) Admin
● Admin can VIEW the User list.
● Admin can ADD/VIEW/UPDATE/DELETE Category.
● Admin can ADD/VIEW/UPDATE/DELETE Sub-Category.
● Admin can ADD/VIEW/UPDATE/DELETE Policy.
● Admin can VIEW/ACTIVATE/CANCEL the buyers’ policy request.
2) Customer
● Customers can VIEW the list of Categories.
● Customers can VIEW the list of Sub-Categories.
● Customers can Apply for a Policy.
● Customer can VIEW the list of Policy he/she holds.